Split native Zenmap saved scan action helpers
